added main mixins
